Add team names to list submission/ export for cryodex

+++ src/routes/todashboard/[tournamentId]/components/TournamentParticipants.svelte
@@ -5,13 +5,18 @@
 	import { nanoid } from "nanoid";
 	import type { Participant } from "$lib/generated/prisma/browser";
 	import { dev } from "$app/environment";
+	import SetTeamDialog from "./SetTeamDialog.svelte";
+	import { exportRowsToCSV } from "$lib/utilities/export";
+	import type { MenuItem } from "$lib/generic/types";
 
 	type Props = {
 		fixedEra: boolean;
 		participants: Participant[];
+		teams: boolean;
 	};
 
-	let { participants, fixedEra }: Props = $props();
+	let { participants, fixedEra, teams }: Props = $props();
+	let setTeamDialogBinding = $state<SetTeamDialog>();
 
 	async function downloadList(id: string) {
 		const result = await getParticipantsGameList(id);
@@ -25,11 +30,49 @@
 			toastController.addToast("Invalid list recieved. Please try again.");
 		}
 	}
+
+	function exportPlayersToCryodex() {
+		const rows: string[] = ['"Name","First Name","Last Name"'];
+		for (const participant of participants) rows.push(`"","${participant.name}",""`);
+		exportRowsToCSV(rows);
+	}
+
+	function exportTeamsToCryodex() {
+		const rows: string[] = ['"Name","First Name","Last Name"'];
+		for (const team of [...new Set(participants.map((p) => p.teamName).filter((t) => t != null))]) rows.push(`"","${team}",""`);
+		exportRowsToCSV(rows);
+	}
+
+	const exportMenuItems = $derived.by(() => {
+		const menuItems: MenuItem[] = [
+			{
+				type: "item",
+				label: "Export Players to Cryodex",
+				onSelect: () => exportPlayersToCryodex()
+			}
+		];
+		if (teams)
+			menuItems.push({
+				type: "item",
+				label: "Export Teams to Cryodex",
+				onSelect: () => exportTeamsToCryodex()
+			});
+		return menuItems;
+	});
 </script>
 
 <section class="card">
-	<p>Participants: <span class="primary">{participants.length}</span></p>
-	<p class="muted">Participants names and emails will be anonymized a few days after the tournament.</p>
+	<div class="flex-between">
+		<div>
+			<p>Participants: <span class="primary">{participants.length}</span></p>
+			<p class="muted">Participants names and emails will be anonymized a few days after the tournament.</p>
+		</div>
+		<DropdownMenu items={exportMenuItems}>
+			{#snippet trigger()}
+				Export
+			{/snippet}
+		</DropdownMenu>
+	</div>
 	<table>
 		<thead>
 			<tr>
@@ -40,6 +83,9 @@
 					<th>Era</th>
 				{/if}
 				<th>Faction</th>
+				{#if teams}
+					<th>Team</th>
+				{/if}
 				<th>Approval Status</th>
 				<th></th>
 				<th></th>
@@ -55,6 +101,9 @@
 						<td>{participant.era}</td>
 					{/if}
 					<td>{participant.faction}</td>
+					{#if teams}
+						<td>{participant.teamName ?? "-"}</td>
+					{/if}
 					<td
 						>{#if participant.approved}
 							<p class="primary">Approved</p>
@@ -79,9 +128,19 @@
 								},
 								{
 									type: "item",
+									label: "Set Team",
+									onSelect: () => {
+										setTeamDialogBinding?.open(participant);
+									}
+								},
+								{
+									type: "separator"
+								},
+								{
+									type: "item",
 									label: "Delete Participant",
 									onSelect: () => {
-										deleteParticipant(participant.id);
+										if (confirm(`Are you sure you wish to remove ${participant.name} from the tournament?`)) deleteParticipant(participant.id);
 									}
 								}
 							]}
@@ -92,12 +151,18 @@
 							{/snippet}
 						</DropdownMenu>
 					</td>
-					<td class={{ last: true, warning: participant.fixedUnits != "[]" }}>{participant.fixedUnits != "[]" ? "List was manually fixed" : ""}</td>
+					<td class={{ last: true, warning: participant.fixedUnits != "[]" }}>
+						<p>
+							{participant.fixedUnits != null || participant.fixedBfs != null || participant.addedUnits != null || participant.addedBfs != null ? "List was manually editted" : ""}
+						</p>
+					</td>
 				</tr>
 			{/each}
 		</tbody>
 	</table>
 </section>
+
+<SetTeamDialog bind:this={setTeamDialogBinding} />
 
 <style>
 	section.card {
